f29h85x: misra: MISRA.ETYPE.COMP.CAST.IMPL.WIDER.2012
A composite expression of essential type 'unsigned char' used as
an operand of the operator '|=' has its other operand having
a wider essential type 'unsigned short int'.

Fixes: TIFSMCU-4426

+++ drivers/crypto/pke/pke.c
@@ -1148,7 +1148,7 @@ static uint32_t PKE_countLeadingZeros(uint32_t x)
 
     /* Left shift until Most significant bit doesn become 1 */
 
-    while ((temp_x & (1 << (bit_count - 1))) == 0) {
+    while ((temp_x & (1 << (bit_count - (uint32_t)1))) == 0) {
         temp_x <<= 1;
         lz++;
     }
